Base `AvatarLogo` on MUI's `Avatar` component
