+++ src/components/AvatarLogo/AvatarLogo.module.scss
@@ -1,7 +1,3 @@
 .avatar_logo {
-  width: 50px;
-  height: 50px;
-  border: 4px transparent solid;
-  border-radius: 50px;
-  overflow: hidden;
+  display: inline-block !important;
 }
